Share pricing/subscriptions query options via queryOptions()
Extract the pricing-page and subscriptions query definitions into
queries.ts, wrapped in `queryOptions<T>()` so the response type flows
through to `prefetchQuery`, `fetchQuery`, `useQuery`, and
`useSuspenseQuery`. Drops the now-redundant explicit `<T>` annotations
on `usePlans`, `useSubscriptions`, `getIdentities`, and the PricingPage
subscriptions query. Addresses review feedback on #2483.
