+++ packages/plugin-zuplo-monetization/src/hooks/usePlans.ts
@@ -1,15 +1,8 @@
-import { useAuth, useZudoku } from "zudoku/hooks";
+import { useZudoku } from "zudoku/hooks";
 import { useSuspenseQuery } from "zudoku/react-query";
-import type { Plan } from "../types/PlanType";
-import { useDeploymentName } from "./useDeploymentName";
+import { pricingPageQuery } from "../queries.js";
 
 export const usePlans = () => {
   const zudoku = useZudoku();
-  const auth = useAuth();
-  const deploymentName = useDeploymentName();
-
-  return useSuspenseQuery<{ items: Plan[] }>({
-    queryKey: [`/v3/zudoku-metering/${deploymentName}/pricing-page`],
-    meta: { context: auth.isAuthenticated ? zudoku : undefined },
-  });
+  return useSuspenseQuery(pricingPageQuery(zudoku));
 };
